+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
@@ -6,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+
+- On zoom to a unit with no location zoom to show subordinates instead.
 - Add button to the unit panel for setting unit location by clicking on the map.
 - Zoom to multiple features using z keyboard shortcut or the zoom button on the feature details panel.
 - Select multiple scenario features with shift+click on the features list, or by shift+click on the map.